requirements.txt
PyQt6==6.11.0
brotli==1.2.0
easyocr==1.7.2
fonttools==4.62.1
numpy==2.4.4
opencv-python-headless==4.13.0.92
Pillow==12.2.0
pytesseract==0.3.13
pyinstaller==6.20.0
